Class_Activity_2/1050090218/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
requests
selenium
fpdf2